+++ Demo/Keyboard/DemoLayoutService.swift
@@ -12,8 +12,8 @@ import KeyboardKit
 import KeyboardKitPro
 #endif
 
-/// This service inherits the standard service, then makes a
-/// few demo-specific adjustments to the standard layout.
+/// This service inherits the standard service, and can then
+/// add one extra button before the spacebar.
 class DemoLayoutService: KeyboardLayout.StandardService {
 
     init(extraKey: ExtraKey) {
@@ -25,8 +25,10 @@ class DemoLayoutService: KeyboardLayout.StandardService {
     enum ExtraKey {
         case none
         case emojiIfNeeded
-        case rocket
+        case keyboardSwitcher
         case localeSwitcher
+        case rocket
+        case url(String)
     }
 
     /// Insert a locale switcher action or a rocket button.
@@ -34,9 +36,11 @@ class DemoLayoutService: KeyboardLayout.StandardService {
         let layout = super.keyboardLayout(for: context)
         switch extraKey {
         case .none: break
-        case .emojiIfNeeded: layout.tryInsertEmojiButton()
-        case .rocket: layout.tryInsertRocketButton()
-        case .localeSwitcher: layout.tryInsertLocaleSwitcher()
+        case .emojiIfNeeded: layout.tryInsert(.keyboardType(.emojis))
+        case .keyboardSwitcher: layout.tryInsert(.nextKeyboard)
+        case .localeSwitcher: layout.tryInsert(.nextLocale)
+        case .rocket: layout.tryInsert(.rocket)
+        case .url(let string): layout.tryInsert(.url(.init(string: string), id: nil))
         }
         return layout
     }
@@ -44,21 +48,8 @@ class DemoLayoutService: KeyboardLayout.StandardService {
 
 private extension KeyboardLayout {
 
-    func tryInsertEmojiButton() {
-        guard let row = bottomRow else { return }
-        let hasEmoji = row.contains(where: { $0.action == .keyboardType(.emojis) })
-        if hasEmoji { return }
-        guard let button = tryCreateBottomRowItem(for: .keyboardType(.emojis)) else { return }
-        itemRows.insert(button, after: .space, atRow: bottomRowIndex)
-    }
-
-    func tryInsertLocaleSwitcher() {
-        guard let item = tryCreateBottomRowItem(for: .nextLocale) else { return }
-        itemRows.insert(item, after: .space, atRow: bottomRowIndex)
-    }
-
-    func tryInsertRocketButton() {
-        guard let button = tryCreateBottomRowItem(for: .rocket) else { return }
-        itemRows.insert(button, before: .space, atRow: bottomRowIndex)
+    func tryInsert(_ action: KeyboardAction) {
+        guard let item = tryCreateBottomRowItem(for: action) else { return }
+        itemRows.insert(item, before: .space, atRow: bottomRowIndex)
     }
 }