+++ Sources/KeyboardKit/Gestures/Gestures+KeyboardButtonGestures.swift
@@ -10,18 +10,19 @@
 import SwiftUI
 
 extension Gestures {
-    
+
     /// This view applies keyboard gestures to any view.
     struct KeyboardButtonGestures<Content: View>: View {
-        
+
         /// Apply a set of action gestures to a view.
         ///
         /// - Parameters:
         ///   - view: The view to apply the gestures to.
         ///   - action: The keyboard action to trigger.
+        ///   - repeatGestureTimer: The repeat gesture timer to use, if any.
         ///   - calloutContext: The callout context to affect, if any.
         ///   - isPressed: An optional binding that can be used to observe the button pressed state.
-        ///   - isInScrollView: Whether or not the gestures are used in a scroll view.
+        ///   - scrollState: The scroll state to use, if any.
         ///   - releaseOutsideTolerance: The percentage of the button size outside its bounds that should count as a release, by default `1.0`.
         ///   - doubleTapAction: The action to trigger when the button is double tapped.
         ///   - longPressAction: The action to trigger when the button is long pressed.
@@ -33,9 +34,10 @@ extension Gestures {
         init(
             view: Content,
             action: KeyboardAction?,
+            repeatGestureTimer: GestureButtonTimer?,
             calloutContext: CalloutContext?,
             isPressed: Binding<Bool>,
-            isInScrollView: Bool,
+            scrollState: GestureButtonScrollState?,
             releaseOutsideTolerance: Double? = nil,
             doubleTapAction: KeyboardGestureAction?,
             longPressAction: KeyboardGestureAction?,
@@ -47,9 +49,10 @@ extension Gestures {
         ) {
             self.view = view
             self.action = action
+            self.repeatGestureTimer = repeatGestureTimer
             self.calloutContext = calloutContext
             self.isPressed = isPressed
-            self.isInScrollView = isInScrollView
+            self.scrollState = scrollState
             self.releaseOutsideTolerance = releaseOutsideTolerance ?? 1.0
             self.doubleTapAction = doubleTapAction
             self.longPressAction = longPressAction
@@ -59,12 +62,13 @@ extension Gestures {
             self.dragAction = dragAction
             self.endAction = endAction
         }
-        
+
         private let view: Content
         private let action: KeyboardAction?
+        private let repeatGestureTimer: GestureButtonTimer?
         private let calloutContext: CalloutContext?
         private let isPressed: Binding<Bool>
-        private let isInScrollView: Bool
+        private let scrollState: GestureButtonScrollState?
         private let releaseOutsideTolerance: Double
         private let doubleTapAction: KeyboardGestureAction?
         private let longPressAction: KeyboardGestureAction?
@@ -73,13 +77,13 @@ extension Gestures {
         private let repeatAction: KeyboardGestureAction?
         private let dragAction: KeyboardDragGestureAction?
         private let endAction: KeyboardGestureAction?
-        
+
         @State
         private var lastDragValue: DragGesture.Value?
-        
+
         @State
         private var shouldApplyReleaseAction = true
-        
+
         var body: some View {
             view.overlay(
                 GeometryReader { geo in
@@ -106,38 +110,16 @@ private extension View {
 
 private extension Gestures.KeyboardButtonGestures {
 
-    @ViewBuilder
     func button(for geo: GeometryProxy) -> some View {
-        if isInScrollView {
-            scrollButton(for: geo)
-        } else {
-            gestureButton(for: geo)
-        }
-    }
-
-    func gestureButton(for geo: GeometryProxy) -> some View {
-        Gestures.GestureButton(
+        GestureButton(
             isPressed: isPressed,
+            scrollState: scrollState,
             pressAction: { handlePress(in: geo) },
             releaseInsideAction: { handleReleaseInside(in: geo) },
             releaseOutsideAction: { handleReleaseOutside(in: geo) },
             longPressAction: { handleLongPress(in: geo) },
             doubleTapAction: { handleDoubleTap(in: geo) },
-            repeatAction: { handleRepeat(in: geo) },
-            dragAction: { handleDrag(in: geo, value: $0) },
-            endAction: { handleGestureEnded(in: geo) },
-            label: { _ in Color.clearInteractable }
-        )
-    }
-
-    func scrollButton(for geo: GeometryProxy) -> some View {
-        Gestures.ScrollViewGestureButton(
-            isPressed: isPressed,
-            pressAction: { handlePress(in: geo) },
-            releaseInsideAction: { handleReleaseInside(in: geo) },
-            releaseOutsideAction: { handleReleaseOutside(in: geo) },
-            longPressAction: { handleLongPress(in: geo) },
-            doubleTapAction: { handleDoubleTap(in: geo) },
+            repeatTimer: repeatGestureTimer,
             repeatAction: { handleRepeat(in: geo) },
             dragAction: { handleDrag(in: geo, value: $0) },
             endAction: { handleGestureEnded(in: geo) },